+++ Core/Inc/hid_mouse.h
@@ -10,7 +10,7 @@
 
 int8_t hid_mouse_move(int8_t x, int8_t y, int8_t wheel);
 int8_t hid_mouse_move_with_pan(int8_t x, int8_t y, int8_t wheel, int8_t pan);
-int8_t hid_mouse_button(uint8_t button, uint8_t mode);
+int8_t hid_mouse_button(uint16_t button, uint8_t mode);
 int8_t hid_mouse_release_all(void);
 
 #endif
